nowe pliki w folderze src

+++ src/main/java/natalia/koc/sklepZoologiczny/repositories/UserRepository.java
@@ -3,8 +3,7 @@ package natalia.koc.sklepZoologiczny.repositories;
 import natalia.koc.sklepZoologiczny.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
     User findUserByUsername(String username);
 }
